# docs/topics/gamification_openapi_2.self-contained.yaml
...
openapi: "3.0.3"
info:
  title: "gamification API"
  description: "gamification API"
  version: "1.0.0"
servers:
  - url: "https://api.tc-syd-02.insa-lyon.fr"
    description: "Production server"
paths:
...
  /api/admin/teachers:
    get:
      summary: "Get all teachers"
      description: "Get all teachers. Returns a list of TeacherUser. Only available to admins - Requires authentication"
      security:
        - cookieAuth: [ ]
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeacherUser"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
                  redirect_url:
                    type: string
                    example: "FRONTEND_URL/login"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Forbidden"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /api/admin/teachers/{id}:
    delete:
      summary: "Remove teacher status from user"
      description: "Remove teacher status from user. Only available to admins - Requires authentication"
      security:
        - cookieAuth: [ ]
      parameters:
        - name: "id"
          in: "path"
          schema:
            type: string
          description: The user id
          required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Teacher deleted succesfully"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
                  redirect_url:
                    type: string
                    example: "FRONTEND_URL/login"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Forbidden"
        "404":
          description: "Not Found"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User not found"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /api/admin/setTps/{id}:
    post:
      summary: "Set teacher's tps"
      description: "Set teacher's tps. Only available to admins - Requires authentication"
      parameters:
        - name: "id"
          in: "path"
          schema:
            type: string
          description: The user id
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                tps:
                  type: array
                  items:
                    type: string
                  example: [ "scrapping", "kafka" ]
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "TPs set successfully"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "TPs not provided"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
                  redirect_url:
                    type: string
                    example: "FRONTEND_URL/login"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "You are not allowed to access this resource"
        "404":
          description: "Not Found"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User not found"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /api/admin/users:
    post:
      summary: "Get users for autocomplete"
      description: "Get users for autocomplete from searchString. Only available to admins - Requires authentication"
      security:
        - cookieAuth: [ ]
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeacherUser"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
                  redirect_url:
                    type: string
                    example: "FRONTEND_URL/login"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "You are not allowed to access this resource"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
  /api/admin/addTeacher:
    post:
      summary: "Add teacher status to user"
      description: "Add teacher status to user. Only available to admins - Requires authentication"
      security:
        - cookieAuth: [ ]
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: object
                properties:
                  smessage:
                    type: string
                    example: "Teacher added succesfully"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized"
                  redirect_url:
                    type: string
                    example: "FRONTEND_URL/login"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "You are not allowed to access this resource"
        "404":
          description: "Not Found"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User not found"
        "409":
          description: "Conflict"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User is already a teacher"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal Server Error"
...
components:
  schemas:
...
    TeacherUser:
      description: "A teacher user"
      type: "object"
      properties:
        id:
          description: "The user's id"
          type: "string"
        name:
          description: "The user's name"
          type: "string"
        surname:
          description: "The user's surname"
          type: "string"
        type:
          description: "The user's type"
          type: "string"
          enum:
            - "student"
            - "teacher"
            - "admin"
        tps:
          description: "The user's available tps"
          type: "array"
          items:
            type: "string"
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid